+++ aws/eks-service/main.tf
@@ -57,7 +57,7 @@ data "aws_iam_policy_document" "this" {
   }
 }
 
-resource "kubernetes_namespace" "this" {
+resource "kubernetes_namespace_v1 " "this" {
   count = try(data.kubernetes_namespace_v1.existing.metadata.0.name) != var.service.namespace || var.create_kubernetes_namespace ? 1 : 0
 
   metadata {
@@ -66,7 +66,7 @@ resource "kubernetes_namespace" "this" {
 }
 
 resource "kubernetes_config_map" "this" {
-  depends_on = [kubernetes_namespace.this]
+  depends_on = [kubernetes_namespace_v1.this]
   metadata {
     name      = var.config_map_env_var_name != null ? var.config_map_env_var_name : "${var.service.name}-config-map"
     namespace = var.service.namespace
@@ -76,7 +76,7 @@ resource "kubernetes_config_map" "this" {
 }
 
 resource "kubernetes_secret" "this" {
-  depends_on = [kubernetes_namespace.this]
+  depends_on = [kubernetes_namespace_v1.this]
   metadata {
     name      = var.secret_env_var_name != null ? var.secret_env_var_name : "${var.service.name}-env-var"
     namespace = var.service.namespace
@@ -113,7 +113,7 @@ resource "aws_iam_role_policy_attachment" "this" {
 }
 
 resource "kubernetes_service_account_v1" "this" {
-  depends_on = [kubernetes_namespace.this]
+  depends_on = [kubernetes_namespace_v1.this]
 
   count = var.service.create_service_account ? 1 : 0
   metadata {
@@ -126,7 +126,7 @@ resource "kubernetes_service_account_v1" "this" {
 }
 
 resource "kubernetes_annotations" "default" {
-  depends_on = [kubernetes_namespace.this]
+  depends_on = [kubernetes_namespace_v1.this]
 
   count       = var.service.create_service_account ? 0 : 1
   api_version = "v1"