+++ aws/eks-helm/metrics-server/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
 ## [0.1.21]() (2024-12-24)
+
 ### Features
+
 * Init module with all the code